_realstate_ragas/requirements.txt
# FastAPI y servidor
fastapi==0.116.1
uvicorn[standard]==0.35.0
starlette==0.47.3
anyio==4.10.0

# Pydantic para validación de datos
pydantic==2.11.9

# Cliente HTTP para validación de URL
httpx==0.28.1

# Dependencias de la API de Google Drive
google-api-python-client==2.178.0
google-auth>=2.47.0,<3.0.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
google-api-core==2.25.1
google-cloud-storage==2.17.0 

# Google AI Integration
langchain-google-genai==2.0.10
google-generativeai==0.8.5

# Operaciones de archivos asíncronos
aiofiles==24.1.0

# Utilities
python-dotenv==1.1.1
numpy==2.2.1

# LangChain Core
langchain==0.3.27
langchain-core==0.3.78
langchain-experimental==0.3.4
langchain-text-splitters==0.3.11
langchain-community==0.3.29

# LangGraph for RAG orchestration
langgraph==0.2.59

# Vector Store
faiss-cpu==1.12.0

# Document Processing
unstructured==0.18.3
markdown==3.9

# Tokenization
tiktoken==0.8.0

# PDF Processing
pypdf==6.5.0

# Testing (pytest y dependencias)
pytest==9.0.2
pytest-cov==7.0.0
coverage==7.13.1
iniconfig==2.3.0
pluggy==1.6.0

# RAGAS for RAG Evaluation 
ragas==0.2.14
mlflow==3.11.1
pandas==2.2.3

# LangSmith for Observability
langsmith==0.5.2

# Additional LangGraph Dependencies
langgraph-checkpoint==2.1.2
langgraph-sdk==0.1.74

# HTTP and Async
httpcore==1.0.9
httpx-sse==0.4.3
aiohttp==3.13.2
aiohappyeyeballs==2.6.1
aiosignal==1.4.0

# Data Processing
datasets==4.4.2
huggingface_hub==1.3.1
pyarrow==22.0.0

# NLP and Text Processing
nltk==3.9.2
langdetect==1.0.9
RapidFuzz==3.14.3

# Serialization and Data
orjson==3.11.5
ormsgpack==1.12.1
jsonpatch==1.33
jsonpointer==3.0.0

# Additional Utilities
tenacity==9.1.2
backoff==2.2.1
diskcache==5.6.3
joblib==1.5.3
tqdm==4.67.1
psutil==7.2.1

# Type Checking and Validation
pydantic-settings==2.12.0
typing-inspect==0.9.0
typing-inspection==0.4.2
dataclasses-json==0.6.7

# File Processing
filetype==1.2.0
python-magic==0.4.27
python-oxmsg==0.0.2
olefile==0.47
beautifulsoup4==4.14.3
lxml==6.0.2
html5lib==1.1

# Compression and Hashing
zstandard==0.25.0
xxhash==3.6.0

# CLI and Terminal
typer-slim==0.21.1
shellingham==1.5.4
click==8.3.1
colorama==0.4.6

# Additional Dependencies
emoji==2.15.0
regex==2025.11.3
python-iso639==2025.11.16
uuid_utils==0.12.0
nest-asyncio==1.6.0
greenlet==3.3.0

# New for Reranking Service
sentence-transformers==5.2.2

# New for MM Rag
pymupdf>=1.23.0
google-genai==1.65.0
